+++ ShooterGame/Source/ShooterGame/Private/Online/ShooterPlayerState.cpp
@@ -1,7 +1,7 @@
 // Copyright 1998-2018 Epic Games, Inc. All Rights Reserved.
 
+#include "Online/ShooterPlayerState.h"
 #include "ShooterGame.h"
-#include "ShooterPlayerState.h"
 
 AShooterPlayerState::AShooterPlayerState(const FObjectInitializer& ObjectInitializer) : Super(ObjectInitializer)
 {
@@ -28,7 +28,7 @@ void AShooterPlayerState::Reset()
 
 void AShooterPlayerState::UnregisterPlayerWithSession()
 {
-	if (!bFromPreviousLevel)
+	if (!IsFromPreviousLevel())
 	{
 		Super::UnregisterPlayerWithSession();
 	}
@@ -120,7 +120,7 @@ int32 AShooterPlayerState::GetDeaths() const
 
 float AShooterPlayerState::GetScore() const
 {
-	return Score;
+	return APlayerState::GetScore();
 }
 
 int32 AShooterPlayerState::GetNumBulletsFired() const
@@ -163,13 +163,15 @@ void AShooterPlayerState::ScorePoints(int32 Points)
 		MyGameState->TeamScores[TeamNumber] += Points;
 	}
 
-	Score += Points;
+    //Score += Points; DEPRECATED
+    int32 CurrentScore = APlayerState::GetScore();
+    APlayerState::SetScore(CurrentScore + Points);
 }
 
 void AShooterPlayerState::InformAboutKill_Implementation(class AShooterPlayerState* KillerPlayerState, const UDamageType* KillerDamageType, class AShooterPlayerState* KilledPlayerState)
 {
 	//id can be null for bots
-	if (KillerPlayerState->UniqueId.IsValid())
+	if (KillerPlayerState->GetUniqueId().IsValid())
 	{	
 		//search for the actual killer before calling OnKill()	
 		for (FConstPlayerControllerIterator It = GetWorld()->GetPlayerControllerIterator(); It; ++It)
@@ -180,7 +182,7 @@ void AShooterPlayerState::InformAboutKill_Implementation(class AShooterPlayerSta
 				// a local player might not have an ID if it was created with CreateDebugPlayer.
 				ULocalPlayer* LocalPlayer = Cast<ULocalPlayer>(TestPC->Player);
 				FUniqueNetIdRepl LocalID = LocalPlayer->GetCachedUniqueNetId();
-				if (LocalID.IsValid() &&  *LocalPlayer->GetCachedUniqueNetId() == *KillerPlayerState->UniqueId)
+				if (LocalID.IsValid() &&  *LocalPlayer->GetCachedUniqueNetId() == *KillerPlayerState->GetUniqueId())
 				{			
 					TestPC->OnKill();
 				}