+++ ShooterGame/Source/ShooterGame/Private/Player/ShooterCheatManager.cpp
@@ -1,7 +1,7 @@
 // Copyright 1998-2018 Epic Games, Inc. All Rights Reserved.
 
-#include "ShooterGame.h"
 #include "Player/ShooterCheatManager.h"
+#include "ShooterGame.h"
 #include "Online/ShooterPlayerState.h"
 #include "Bots/ShooterAIController.h"
 
@@ -30,7 +30,7 @@ void UShooterCheatManager::ToggleMatchTimer()
 	AShooterPlayerController* MyPC = GetOuterAShooterPlayerController();
 
 	AShooterGameState* const MyGameState = MyPC->GetWorld()->GetGameState<AShooterGameState>();
-	if (MyGameState && MyGameState->Role == ROLE_Authority)
+	if (MyGameState && MyGameState->GetLocalRole() == ROLE_Authority)
 	{
 		MyGameState->bTimerPaused = !MyGameState->bTimerPaused;
 		MyPC->ClientMessage(FString::Printf(TEXT("Match timer: %s"), MyGameState->bTimerPaused ? TEXT("PAUSED") : TEXT("running")));
@@ -53,7 +53,7 @@ void UShooterCheatManager::ChangeTeam(int32 NewTeamNumber)
 	AShooterPlayerController* MyPC = GetOuterAShooterPlayerController();
 
 	AShooterPlayerState* MyPlayerState = Cast<AShooterPlayerState>(MyPC->PlayerState);
-	if (MyPlayerState && MyPlayerState->Role == ROLE_Authority)
+	if (MyPlayerState && MyPlayerState->GetLocalRole() == ROLE_Authority)
 	{
 		MyPlayerState->SetTeamNum(NewTeamNumber);
 		MyPC->ClientMessage(FString::Printf(TEXT("Team changed to: %d"), MyPlayerState->GetTeamNum()));