+++ ShooterGame/Source/ShooterGame/Private/ShooterTeamStart.cpp
@@ -1,7 +1,7 @@
 // Copyright 1998-2018 Epic Games, Inc. All Rights Reserved.
 
-#include "ShooterGame.h"
 #include "ShooterTeamStart.h"
+#include "ShooterGame.h"
 
 AShooterTeamStart::AShooterTeamStart(const FObjectInitializer& ObjectInitializer) 
 	: Super(ObjectInitializer)