+++ ShooterGame/Source/ShooterGame/Private/UI/Menu/Widgets/SShooterLeaderboard.cpp
@@ -1,7 +1,7 @@
 // Copyright 1998-2018 Epic Games, Inc. All Rights Reserved.
 
-#include "ShooterGame.h"
 #include "SShooterLeaderboard.h"
+#include "ShooterGame.h"
 #include "ShooterStyle.h"
 #include "ShooterUIHelpers.h"