+++ ShooterGame/Source/ShooterGame/Private/UI/ShooterHUD.cpp
@@ -1,7 +1,7 @@
 // Copyright 1998-2018 Epic Games, Inc. All Rights Reserved.
 
-#include "ShooterGame.h"
 #include "UI/ShooterHUD.h"
+#include "ShooterGame.h"
 #include "SShooterScoreboardWidget.h"
 #include "SChatWidget.h"
 #include "Engine/ViewportSplitScreen.h"