+++ ShooterGame/Source/ShooterGame/Private/UI/Style/ShooterMenuWidgetStyle.cpp
@@ -1,7 +1,7 @@
 // Copyright 1998-2018 Epic Games, Inc. All Rights Reserved.
 
-#include "ShooterGame.h"
 #include "ShooterMenuWidgetStyle.h"
+#include "ShooterGame.h"
 
 FShooterMenuStyle::FShooterMenuStyle()
 {