+++ ShooterGame/Source/ShooterGame/Private/Weapons/ShooterWeapon_Projectile.cpp
@@ -1,7 +1,7 @@
 // Copyright 1998-2018 Epic Games, Inc. All Rights Reserved.
 
-#include "ShooterGame.h"
 #include "Weapons/ShooterWeapon_Projectile.h"
+#include "ShooterGame.h"
 #include "Weapons/ShooterProjectile.h"
 
 AShooterWeapon_Projectile::AShooterWeapon_Projectile(const FObjectInitializer& ObjectInitializer) : Super(ObjectInitializer)
@@ -75,7 +75,7 @@ void AShooterWeapon_Projectile::ServerFireProjectile_Implementation(FVector Orig
 	AShooterProjectile* Projectile = Cast<AShooterProjectile>(UGameplayStatics::BeginDeferredActorSpawnFromClass(this, ProjectileConfig.ProjectileClass, SpawnTM));
 	if (Projectile)
 	{
-		Projectile->Instigator = Instigator;
+		Projectile->SetInstigator(GetInstigator());
 		Projectile->SetOwner(this);
 		Projectile->InitVelocity(ShootDir);